Update for SEIR129
Removed old references to body-parser since Express has that baked in
for nearly 3 years now.  Updated and streamlined the introductory text.
Fixed all of the references to the 'internet' that should be 'Web' as
they are different and that's a common interview question, so we should
not reinforce incorrect usage of the terms even if they are often used
interchangeably.